Update tools_v2.0.sh
The updated version has each function added for the step by step guide for each of the tools and it makes it easier to update the documentation in case of any updates in the tools. Further, I have added the display of asterisks in order to make the display uncluttered and easy to understand for the user.