Addition Of the AI insights for the new tools, which was not there in the version 1.0 of AI insights.
I have added the ai insights functionality for the tools, in pen test and SCR, I did it for bandit, OWASP ZAP, John the ripper, SQLmap,legion, and SonarQUBE. Which was not present in the previous version of the tool. Further, I have also changed the function of generating AI insights to ignore any special characters that might cause an error while passing the payload during the API calls. Further, I have enhanced the display of the insights by the addition of asterisks and new colours to make it more understandable and readable.